Rename server dir
