+++ server/test/nfc_reader.py
